FIx backend

+++ controllers/userController.js
@@ -8,9 +8,7 @@ export const getCurrentUser = async (req, res) => {
   const user = await User.findOne({ _id: req.user.userId });
 
   console.log('current user: ', user);
-  if (!user) {
-    throw new UnauthenticatedError('invalid credentials');
-  }
+  if (!user) throw new UnauthenticatedError('invalid credentials');
 
   const userWithoutPassword = user.toJSON();
 
@@ -34,13 +32,10 @@ export const updateUser = async (req, res) => {
 
   const updatedUser = await User.findByIdAndUpdate(req.user.userId, req.body);
 
-  if (cvFile && updatedUser.cv) {
-    await fs.unlink(`public${updatedUser.cv}`);
-  }
+  if (cvFile && updatedUser.cv) await fs.unlink(`public${updatedUser.cv}`);
 
-  if (avatarFile && updatedUser.avatar) {
+  if (avatarFile && updatedUser.avatar)
     await fs.unlink(`public${updatedUser.avatar}`);
-  }
 
   res.status(StatusCodes.OK).json({ msg: 'user updated' });
 };
